Update on "[quantization] Factored out the default mappings"
Differential Revision: [D17694475](https://our.internmc.facebook.com/intern/diff/D17694475)
